Refactor things out of util with limited use, move more to service
