+++ modules/services/keepRight.js
@@ -9,10 +9,13 @@ import { json as d3_json } from 'd3-request';
 import { request as d3_request } from 'd3-request';
 
 import { geoExtent, geoVecAdd } from '../geo';
-import { services } from './index';
 import { krError } from '../osm';
-
+import { services } from './index';
+import { t } from '../util/locale';
 import { utilRebind, utilTiler, utilQsString } from '../util';
+
+import { errorTypes } from '../../data/keepRight.json';
+
 
 var tiler = utilTiler();
 var dispatch = d3_dispatch('loaded');
@@ -21,6 +24,7 @@ var _krCache;
 var _krZoom = 14;
 var apibase = 'https://www.keepright.at/';
 var defaultRuleset = [0,30,40,50,70,90,100,110,120,130,150,160,170,180,191,192,193,194,195,196,197,198,201,202,203,204,205,206,207,208,210,220,231,232,270,281,282,283,284,285,291,292,293,294,295,296,297,298,311,312,313,320,350,370,380,401,402,411,412,413];
+
 
 function abortRequest(i) {
     if (i) {
@@ -58,6 +62,179 @@ function updateRtree(item, replace) {
 }
 
 
+function tokenReplacements(datum) {
+    if (!(datum instanceof krError)) return;
+
+    var replacements = {};
+    var html_re = new RegExp(/<\/[a-z][\s\S]*>/);
+    var commonEntities = ['node', 'way', 'relation', 'highway', 'cycleway', 'waterway', 'riverbank'];
+
+    var errorType;
+    var errorTemplate;
+    var errorDescription;
+    var errorRegex;
+    var errorMatch;
+
+    // find the matching template from the error schema
+    errorType = '_' + datum.error_type;
+    errorTemplate = errorTypes.errors[errorType] || errorTypes.warnings[errorType];
+    if (!errorTemplate) return;
+
+    // some descriptions are just fixed text
+    if (!('regex' in errorTemplate)) return;
+
+    // regex pattern should match description with variable details captured as groups
+    errorDescription = datum.description;
+    errorRegex = new RegExp(errorTemplate.description);
+    errorMatch = errorRegex.exec(errorDescription);
+    if (!errorMatch) {
+        // TODO: Remove, for regex dev testing
+        console.log('Unmatched:', errorType, errorDescription, errorRegex);
+        return;
+    }
+
+    errorMatch.forEach(function(group, index) {
+        var idType;
+
+        // index 0 is the whole match, skip it
+        if (!index) return;
+
+        // link IDs if present in the group
+        idType = 'IDs' in errorTemplate ? errorTemplate.IDs[index-1] : '';
+        if (idType && group) {
+            group = parseError(group, idType);
+        } else if (html_re.test(group)) {
+            // escape any html in non-IDs
+            group = '\\' +  group + '\\';
+        }
+
+        // translate common words (e.g. node, way, relation)
+        if (commonEntities.includes(group)) {
+            group = t('QA.keepRight.entities.' + group);
+        }
+
+        replacements['var' + index] = group;
+    });
+
+    return replacements;
+}
+
+
+function parseError(group, idType) {
+
+    function fillPlaceholder(d) { return '<span><a class="kr_error_description-id">' + d + '</a></span>'; }
+
+    // arbitrary node list of form: #ID, #ID, #ID...
+    function parseError211(list) {
+        var newList = [];
+        var items = list.split(', ');
+
+        items.forEach(function(item) {
+            // ID has # at the front
+            var id = fillPlaceholder('n' + item.slice(1));
+            newList.push(id);
+        });
+
+        return newList.join(', ');
+    }
+
+    // arbitrary way list of form: #ID(layer),#ID(layer),#ID(layer)...
+    function parseError231(list) {
+        var newList = [];
+        var items = list.split(',');
+
+        items.forEach(function(item) {
+            var id;
+            var layer;
+
+            // item of form "#ID(layer)"
+            item = item.split('(');
+
+            // ID has # at the front
+            id = item[0].slice(1);
+            id = fillPlaceholder('w' + id);
+
+            // layer has trailing )
+            layer = item[1].slice(0,-1);
+
+            // TODO: translation
+            newList.push(id + ' (layer: ' + layer + ')');
+        });
+
+        return newList.join(', ');
+    }
+
+    // arbitrary node/relation list of form: from node #ID,to relation #ID,to node #ID...
+    function parseError294(list) {
+        var newList = [];
+        var items = list.split(',');
+
+        items.forEach(function(item) {
+            var role;
+            var idType;
+            var id;
+
+            // item of form "from/to node/relation #ID"
+            item = item.split(' ');
+
+            // to/from role is more clear in quotes
+            role = '"' + item[0] + '"';
+
+            // first letter of node/relation provides the type
+            idType = item[1].slice(0,1);
+
+            // ID has # at the front
+            id = item[2].slice(1);
+            id = fillPlaceholder(idType + id);
+
+            item = [role, item[1], id].join(' ');
+            newList.push(item);
+        });
+
+        return newList.join(', ');
+    }
+
+    // TODO: Handle error 401 template addition
+
+    // arbitrary node list of form: #ID,#ID,#ID...
+    function parseWarning20(list) {
+        var newList = [];
+        var items = list.split(',');
+
+        items.forEach(function(item) {
+            // ID has # at the front
+            var id = fillPlaceholder('n' + item.slice(1));
+            newList.push(id);
+        });
+
+        return newList.join(', ');
+    }
+
+    switch (idType) {
+        // simple case just needs a linking span
+        case 'n':
+        case 'w':
+        case 'r':
+            group = fillPlaceholder(idType + group);
+            break;
+        // some errors have more complex ID lists/variance
+        case '211':
+            group = parseError211(group);
+            break;
+        case '231':
+            group = parseError231(group);
+            break;
+        case '294':
+            group = parseError294(group);
+            break;
+        case '20':
+            group = parseWarning20(group);
+    }
+
+    return group;
+}
+
+
 export default {
     init: function() {
         if (!_krCache) {
@@ -77,11 +254,8 @@ export default {
 
     // KeepRight API:  http://osm.mueschelsoft.de/keepright/interfacing.php
     loadErrors: function(context, projection) {
-        var options = {
-            format: 'geojson'
-        };
+        var options = { format: 'geojson' };
         var rules = defaultRuleset.join();
-        var that = this;
 
         // determine the needed tiles to cover the view
         var tiles = tiler
@@ -135,6 +309,8 @@ export default {
                             schema: props.schema,
                             title: props.title
                         });
+
+                        d.replacements = tokenReplacements(d);
 
                         _krCache.keepRight[d.id] = d;
                         _krCache.rtree.insert(encodeErrorRtree(d));
