--- modules/util/keepRight/index.js
@@ -1,3 +1,0 @@
-export { errorTypes } from './errorSchema.json';
-export { parseError } from './parse_error';
-export { parseErrorDescriptions, clickLink } from './keepRight_error';
